Move transitions to semantic
