+++ sdk/src/types/block/output/foundry.rs
@@ -23,7 +23,7 @@ use crate::types::block::{
     },
     payload::signed_transaction::{TransactionCapabilities, TransactionCapabilityFlag},
     protocol::{ProtocolParameters, WorkScore, WorkScoreParameters},
-    semantic::{SemanticValidationContext, StateTransitionError, StateTransitionVerifier, TransactionFailureReason},
+    semantic::{SemanticValidationContext, StateTransitionError, TransactionFailureReason},
     unlock::Unlock,
     Error,
 };
@@ -521,81 +521,6 @@ impl WorkScore for FoundryOutput {
 }
 
 impl MinimumOutputAmount for FoundryOutput {}
-
-impl StateTransitionVerifier for FoundryOutput {
-    fn creation(next_state: &Self, context: &SemanticValidationContext<'_>) -> Result<(), StateTransitionError> {
-        let account_chain_id = ChainId::from(*next_state.account_address().account_id());
-
-        if let (Some(Output::Account(input_account)), Some(Output::Account(output_account))) = (
-            context.input_chains.get(&account_chain_id),
-            context.output_chains.get(&account_chain_id),
-        ) {
-            if input_account.foundry_counter() >= next_state.serial_number()
-                || next_state.serial_number() > output_account.foundry_counter()
-            {
-                return Err(StateTransitionError::InconsistentFoundrySerialNumber);
-            }
-        } else {
-            return Err(StateTransitionError::MissingAccountForFoundry);
-        }
-
-        let token_id = next_state.token_id();
-        let output_tokens = context.output_native_tokens.get(&token_id).copied().unwrap_or_default();
-        let TokenScheme::Simple(ref next_token_scheme) = next_state.token_scheme;
-
-        // No native tokens should be referenced prior to the foundry creation.
-        if context.input_native_tokens.contains_key(&token_id) {
-            return Err(StateTransitionError::InconsistentNativeTokensFoundryCreation);
-        }
-
-        if output_tokens != next_token_scheme.minted_tokens() || !next_token_scheme.melted_tokens().is_zero() {
-            return Err(StateTransitionError::InconsistentNativeTokensFoundryCreation);
-        }
-
-        Ok(())
-    }
-
-    fn transition(
-        current_state: &Self,
-        next_state: &Self,
-        context: &SemanticValidationContext<'_>,
-    ) -> Result<(), StateTransitionError> {
-        Self::transition_inner(
-            current_state,
-            next_state,
-            &context.input_native_tokens,
-            &context.output_native_tokens,
-            context.transaction.capabilities(),
-        )
-    }
-
-    fn destruction(current_state: &Self, context: &SemanticValidationContext<'_>) -> Result<(), StateTransitionError> {
-        if !context
-            .transaction
-            .has_capability(TransactionCapabilityFlag::DestroyFoundryOutputs)
-        {
-            return Err(TransactionFailureReason::TransactionCapabilityFoundryDestructionNotAllowed)?;
-        }
-
-        let token_id = current_state.token_id();
-        let input_tokens = context.input_native_tokens.get(&token_id).copied().unwrap_or_default();
-        let TokenScheme::Simple(ref current_token_scheme) = current_state.token_scheme;
-
-        // No native tokens should be referenced after the foundry destruction.
-        if context.output_native_tokens.contains_key(&token_id) {
-            return Err(StateTransitionError::InconsistentNativeTokensFoundryDestruction);
-        }
-
-        // This can't underflow as it is known that minted_tokens >= melted_tokens (syntactic rule).
-        let minted_melted_diff = current_token_scheme.minted_tokens() - current_token_scheme.melted_tokens();
-
-        if minted_melted_diff != input_tokens {
-            return Err(StateTransitionError::InconsistentNativeTokensFoundryDestruction);
-        }
-
-        Ok(())
-    }
-}
 
 impl Packable for FoundryOutput {
     type UnpackError = Error;
